Refactor js code to be cleaner
