+++ viewer/templates/base.html
@@ -25,7 +25,7 @@
     <link rel="stylesheet" type="text/css" href="/static/css/main.css">
 
     <!-- Custom JavaScript-->
-    <script type="text/javascript" src="/static/js/main.js"></script>
+    <script type="module" src="/static/js/main.js"></script>
 </head>
 <body>
